+++ python/taichi/lang/ast/ast_transformer.py
@@ -468,8 +468,8 @@ class ASTTransformer(Builder):
                 # For args[0], it is an ast.Attribute, because it loads the
                 # attribute, |ptr|, of the expression |ret_expr|. Therefore we
                 # only need to replace the object part, i.e. args[0].value
-            return None
-        ctx.return_data = node.value.ptr
+        else:
+            ctx.return_data = node.value.ptr
         if not impl.get_runtime().experimental_real_function:
             ctx.returned = True
         return None
