+++ tests/python/test_return.py
@@ -1,18 +1,7 @@
+import pytest
+
 import taichi as ti
 from taichi import approx
-
-
-@ti.test()
-@ti.must_throw(ti.TaichiSyntaxError)
-def _test_return_not_last_stmt():  # TODO: make this work
-    x = ti.field(ti.i32, ())
-
-    @ti.kernel
-    def kernel() -> ti.i32:
-        return 1
-        x[None] = 233
-
-    kernel()
 
 
 @ti.test()
@@ -65,3 +54,50 @@ def test_binary_func_ret():
     _test_binary_func_ret(ti.f32, ti.i32, ti.f32, float)
     _test_binary_func_ret(ti.i32, ti.f32, ti.i32, int)
     _test_binary_func_ret(ti.f32, ti.i32, ti.i32, int)
+
+
+@ti.test()
+def test_return_in_static_if():
+    @ti.kernel
+    def foo(a: ti.template()) -> ti.i32:
+        if ti.static(a == 1):
+            return 1
+        elif ti.static(a == 2):
+            return 2
+        return 3
+
+    assert foo(1) == 1
+    assert foo(2) == 2
+    assert foo(123) == 3
+
+
+@ti.test()
+def test_func_multiple_return():
+    @ti.func
+    def safe_sqrt(a):
+        if a > 0:
+            return ti.sqrt(a)
+        else:
+            return 0.0
+
+    @ti.kernel
+    def kern(a: float):
+        print(safe_sqrt(a))
+
+    with pytest.raises(
+            ti.TaichiCompilationError,
+            match='Return inside non-static if/for is not supported'):
+        kern(-233)
+
+
+@ti.test()
+def test_return_inside_static_for():
+    @ti.kernel
+    def foo() -> ti.i32:
+        a = 0
+        for i in ti.static(range(10)):
+            a += i * i
+            if ti.static(i == 8):
+                return a
+
+    assert foo() == 204
